Change TLS args in Docker configuration example
The TLS args given in this example were misleading, because they were preventing the Docker daemon to listen on the port 2376 (even if the daemon still succeeds to start...).

The reason is very simple: when using `sudo ros tls gen --server ...`, the server and CA certificates will be stored in `/etc/docker/tls`. But their paths were given as relative paths, and not absolute paths, in the previous example.