Add "proxy_buffering off" to nginx reverse proxy config
Add "proxy_buffering off" to the nginx reverse proxy config example.

By default, proxy buffering is turned on and because of that nginx doesn't send the response until it decides to flush proxy buffers. That causes commands that wait for a response (such as `kubectl logs -f` or `kubectl get pods -w`) to not receive anything and just hang.