Small change to quick-start
This tweaks the big NOTE at the top of the quick-start guide to point to kubernetes tutorials for any k8s newbies.
These tutorials are actually pretty great and explain how to create a deployment and get this exposed to the outside world.

We will look into adding docs around Traefik.