Normalize page weights and initial add storage section
This tweaks the page weights so they make more sense.
We now have these in multiples of ten so we can easily move up or down new sections as we go beetween these numbers.
The Volumes and Storage section as added, intial only.
A separate commit will have the content for the Volumes and Storage page.