Update and centralize cluster capabilities table
Currently there are 4 different occurrences of the cluster capabilities table that compares which features you can use in Rancher with RKE clusters, hosted cloud provider clusters and imported clusters. One of them is very detailed but a bit outdated, the other ones only mention basic features and are not complete.

This commit centralizes the table and includes it on all 4 pages and updates it for features introduced with 2.4

Signed-off-by: Bastian Hofmann <bashofmann@gmail.com>