Fix order of extra_binds
Extra bind mounts have the order <source>:<destination>. When copying the example from the documentation, the kubelet was not coming up anymore because it tried to mount `/host/dev` to `/dev`. Output from `docker inspect` on the kubelet container:
```
            {
                "Type": "bind",
                "Source": "/host/dev",
                "Destination": "/dev",
                "Mode": "",
                "RW": true,
                "Propagation": "rprivate"
            },
```

This commit fixes the order of the paths in the documentation.