Remove note for Auto Upgrade for >2.5.8
Prior to 2.5.8, enabling Auto Upgrade for GKE hosted clusters was not
recommended because it could cause the state of the cluster in GKE to
become out of sync with the state of the cluster in Rancher. With 2.5.8,
this is not a concern any more because of the syncing mechanism in the
new GKE provisioner. Therefore, it's recommended to use the Auto Upgrade
feature, which defaults to enabled in both Rancher and the GKE console.
This change removes the note erroneously added in 74b31fae.