Add "Mutable" indicator for GKE config-reference
The Rancher UI prevents a user from editing a field for a provisioned
GKE cluster that GKE does not support changing, but the Rancher API
schema allows updates to some fields that will get ignored and cause
confusion to API users. This change (for 2.5.8 only) tries to mitigate
that confusion by pointing out which fields can be changed.