Update advanced TLS settings page
This adds a detailed description on how to configure tls version and ciphers when Rancher is installed in a Kubernetes cluster. TLS is offloaded at the ingress controller, which means these settings need to be set there instead of the Rancher Pod/Container.

For running Rancher as a single Docker container, the default values have been updated to the ones currently supported.

Signed-off-by: Bastian Hofmann <bashofmann@gmail.com>